ai-traffic-signal-optimization-main/requirements.txt
flask==3.0.3
flask-cors==4.0.0
ultralytics==8.3.0
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
pymongo==4.10.1
python-dotenv==1.0.1
matplotlib==3.9.0
